🔧 MAINTAIN: Split sphinx extension into separate module (#409)
