+++ myst_nb/__init__.py
@@ -5,7 +5,7 @@ __version__ = "0.14.0"
 def setup(app):
     """Sphinx extension setup."""
     # we import this locally, so sphinx is not automatically imported
-    from .sphinx_ import sphinx_setup
+    from .sphinx_ext import sphinx_setup
 
     return sphinx_setup(app)
